Add dependency:greenDAO

+++ app/src/main/java/com/maxwell/youchat/model/Message.java
@@ -3,28 +3,28 @@ package com.maxwell.youchat.model;
 import java.sql.Timestamp;
 
 public class Message {
-    private Integer id;
-    private Integer sendUserId;
-    private Integer receiveUserId;
+    private Long id;
+    private Long sendUserId;
+    private Long receiveUserId;
     private Timestamp createTime;
     private String content;
 
-    public Message(Integer sendUserId, Integer receiveUserId, Timestamp createTime, String content) {
+    public Message(Long sendUserId, Long receiveUserId, Timestamp createTime, String content) {
         this.sendUserId = sendUserId;
         this.receiveUserId = receiveUserId;
         this.createTime = createTime;
         this.content = content;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public Integer getSendUserId() {
+    public Long getSendUserId() {
         return sendUserId;
     }
 
-    public Integer getReceiveUserId() {
+    public Long getReceiveUserId() {
         return receiveUserId;
     }
